Cap adverse fills, extend 1h window, panel-adjudicated Stage 3, regime exhibit
- HALF_SPREAD_CAP=0.5: extreme listing-minute ranges (high/open > 5x) pushed
  the adverse shift past 100%, producing negative sell fills that FLIPPED
  short-return signs on 116 open-entry bars (a false-REAL vector). Fills now
  bottom out at 50% adverse; non-positive fills return NaN, never P&L.
- 1h klines extended to T0+15.5d: the +24h|14d exit lands exactly at +15d
  and previously force-exited an hour early on every event.
- Stage 3 per docs/DECISIONS.md D1 (3-judge outcome-blind adjudication,
  unanimous): null-calibrated expected-max hurdle governs (per-trial
  variances, seeded MC, N = all trials incl. controls, correlation not
  discounted); paper plug-in disclosed alongside; Bonferroni dependence-free
  cross-check printed.
- In-progress candles filtered from REST ingest; FORWARD_MARGIN_DAYS=16.
- Unreached stages render as '—', not FAIL; PBO row discloses unpurged-CSCV
  bias direction; per-year regime exhibit + walk-forward fold detail in the
  report; REAL-verdict text spells out funding/availability/squeeze-tail
  caveats and forbids reading it as deployment authorization.

+++ validation/events.py
@@ -212,8 +212,8 @@ def ingest_event_klines(conn, progress=print, workers=10):
     for ev in evs:
         t0 = ev["first_trade_time"]
         want = [
-            ("1m", t0, t0 + config.KLINE_1M_HOURS * 3_600_000),
-            ("1h", t0, t0 + config.KLINE_1H_DAYS * 86_400_000),
+            ("1m", t0, t0 + int(config.KLINE_1M_HOURS * 3_600_000)),
+            ("1h", t0, t0 + int(config.KLINE_1H_DAYS * 86_400_000)),
         ]
         for interval, lo, hi in want:
             got = have.get((ev["symbol"], interval))
